feat: changing int to float for more versatility

+++ estudos-linguagens/desafios-python/mini-projeto-calculadora/calculadora.py
@@ -1,7 +1,7 @@
 # Mini projeto de calculadora via terminal com match case
 
-number_one = int(input("Digite um número: "))
-number_two = int(input("Digite outro número: "))
+number_one = float(input("Digite um número: "))
+number_two = float(input("Digite outro número: "))
 
 operation = input("Digite sua operação (Ela pode ser +(mais), -(menos), /(divisão), *(multiplicação)): ")
